Isolate input code from libretro.cpp
