+++ Makefile.common
@@ -219,7 +219,8 @@ ifneq ($(HAVE_GRIFFIN), 1)
                   $(MEDNAFEN_DIR)/mempatcher.cpp \
                   $(MEDNAFEN_DIR)/video/Deinterlacer.cpp \
                   $(MEDNAFEN_DIR)/video/surface.cpp \
-                  $(CORE_DIR)/libretro.cpp
+                  $(CORE_DIR)/libretro.cpp \
+                  $(CORE_DIR)/input.cpp
 
    SOURCES_C +=   $(MEDNAFEN_DIR)/file.c \
                   $(CORE_DIR)/rsx/rsx_lib_soft.c \
